# tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "EnvelopeModel_GlossaryIndex_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GlossaryIndex"
          },
          "links": {
            "$ref": "#/components/schemas/LinksModel"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaModel"
          }
        },
        "required": [
          "data",
          "meta",
          "links"
        ],
        "title": "EnvelopeModel[GlossaryIndex]",
        "type": "object"
      },
      "EnvelopeModel_GlossaryTermDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GlossaryTermDetail"
          },
          "links": {
            "$ref": "#/components/schemas/LinksModel"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaModel"
          }
        },
        "required": [
          "data",
          "meta",
          "links"
        ],
        "title": "EnvelopeModel[GlossaryTermDetail]",
        "type": "object"
      },
...
      "EnvelopeModel_list_GlossaryTerm__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/GlossaryTerm"
            },
            "title": "Data",
            "type": "array"
          },
          "links": {
            "$ref": "#/components/schemas/LinksModel"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaModel"
          }
        },
        "required": [
          "data",
          "meta",
          "links"
        ],
        "title": "EnvelopeModel[list[GlossaryTerm]]",
        "type": "object"
      },
...
      "GlossaryIndex": {
        "properties": {
          "entries": {
            "description": "Surface forms, longest-match first.",
            "items": {
              "$ref": "#/components/schemas/IndexEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "index_version": {
            "description": "Content address of the term rows it was built from.",
            "title": "Index Version",
            "type": "string"
          },
          "stopwords": {
            "description": "Glossary terms excluded from auto-highlighting.",
            "items": {
              "type": "string"
            },
            "title": "Stopwords",
            "type": "array"
          }
        },
        "required": [
          "index_version",
          "entries",
          "stopwords"
        ],
        "title": "GlossaryIndex",
        "type": "object"
      },
      "GlossaryTerm": {
        "properties": {
          "aliases": {
            "description": "Other spellings that resolve to this term.",
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "domain_tags": {
            "description": "Domains the term belongs to.",
            "items": {
              "type": "string"
            },
            "title": "Domain Tags",
            "type": "array"
          },
          "highlightable": {
            "description": "False for ordinary words: reachable, but never auto-highlighted (M7).",
            "title": "Highlightable",
            "type": "boolean"
          },
          "short_definition": {
            "description": "One-line definition for a hover.",
            "title": "Short Definition",
            "type": "string"
          },
          "term": {
            "description": "Surface form of the term.",
            "title": "Term",
            "type": "string"
          },
          "term_id": {
            "description": "Stable id; this is what meta.labels points at.",
            "title": "Term Id",
            "type": "string"
          }
        },
        "required": [
          "term_id",
          "term",
          "aliases",
          "short_definition",
          "domain_tags",
          "highlightable"
        ],
        "title": "GlossaryTerm",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LinksModel": {
        "additionalProperties": true,
        "properties": {
          "explain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pre-built /v1/explain call for every handle in the body.",
            "title": "Explain"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Next page, or null at the end.",
            "title": "Next"
          },
          "self": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Canonical path of this response.",
            "title": "Self"
          }
        },
        "title": "LinksModel",
        "type": "object"
      },
...
      "MetaModel": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "$ref": "#/components/schemas/AsOfModel",
            "description": "Requested and resolved vintage (SB-04 \u00a72.5)."
          },
          "deprecations": {
            "description": "Deprecations in force for this operation.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Deprecations",
            "type": "array"
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "JSON Pointer to glossary term_id (DIR-8).",
            "title": "Labels",
            "type": "object"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor for the next page; null at the end.",
            "title": "Next Cursor"
          },
          "request_id": {
            "description": "ULID for this request; echoed in every problem body.",
            "title": "Request Id",
            "type": "string"
          },
          "source_freshness": {
            "additionalProperties": true,
            "description": "Per-source retrieval vintage and state.",
            "title": "Source Freshness",
            "type": "object"
          },
          "warnings": {
            "description": "Structured warnings: code, detail, optional pointer.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "as_of"
        ],
        "title": "MetaModel",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": true,
        "description": "The wire shape of every failure (RFC 9457).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What went wrong on this request.",
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-field failures, each with a JSON Pointer.",
            "title": "Errors"
          },
          "instance": {
            "description": "Path of the request that failed.",
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "description": "ULID joining this response to the audit trail.",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code, repeated in the body per RFC 9457.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the error code.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Absolute URI of the error code; resolves at /v1/errors/{code}.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance",
          "request_id"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Static owner key for this deployment. /healthz and the frontend are open.",
        "in": "header",
        "name": "X-Glasswell-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Glass-box upstream analytics on public data. Every number this API serves carries a\nderivation handle in band, and `GET /v1/explain` walks that handle back to the\nchecksummed government file it came from.\n\nConventions that hold everywhere: responses are `data` / `meta` / `links`; collections\nare cursor-paginated with no offset parameter; failures are RFC 9457 problem documents\nwhose `type` resolves at `/v1/errors/{code}`; and `as_of` selects knowledge time, with\nthe resolved vintage reported back in `meta.as_of`.\n\nThis deployment is the North Dakota slice: wells, geometry, monthly production, the\nlineage spine, the conformance registry, quarantine and the glossary. Forecasts,\neconomics and other basins are not served.",
    "title": "glasswell",
    "version": "0.1.0",
    "x-glasswell-freeze": {
      "checked_by": "tests/contract/openapi_diff.py",
      "criterion": "S1",
      "frozen_on": "2026-08-20",
      "policy": "Additive change only. A path, operation, parameter, response field or enum value that is published here is published for the life of /v1; removing one, or making an optional request parameter required, is a /v2 event (blueprint \u00a73.6.1).",
      "status": "frozen",
      "surface": "v1"
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/glossary": {
      "get": {
        "description": "Every term the product uses, alphabetically, with its short definition and its aliases. `q` matches the term and its aliases. The glossary is data promoted through the same rules as everything else, not markup embedded in views.",
        "operationId": "list_glossary_terms",
        "parameters": [
          {
            "description": "Opaque cursor from meta.next_cursor. There is no offset parameter.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from meta.next_cursor. There is no offset parameter.",
              "title": "Cursor"
            }
          },
          {
            "description": "Page size, 100 by default, 200 at most (SB-07 \u00a79.5 caps the spine collections lower than the data ones).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Page size, 100 by default, 200 at most (SB-07 \u00a79.5 caps the spine collections lower than the data ones).",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Match a term or one of its aliases.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Match a term or one of its aliases.",
              "title": "Q"
            }
          },
          {
            "description": "Filter to one domain tag.",
            "in": "query",
            "name": "domain_tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one domain tag.",
              "title": "Domain Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeModel_list_GlossaryTerm__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "cursor_malformed": {
                    "summary": "Cursor malformed",
                    "value": {
                      "detail": "The cursor did not decode to the four declared fields.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Cursor malformed",
                      "type": "/v1/errors/cursor_malformed"
                    }
                  },
                  "cursor_query_mismatch": {
                    "summary": "Cursor does not match this query",
                    "value": {
                      "detail": "The cursor was minted against a different filter set. Continuing would return a page from a different result set (SB-04 \u00a72.3).",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Cursor does not match this query",
                      "type": "/v1/errors/cursor_query_mismatch"
                    }
                  },
                  "validation_failed": {
                    "summary": "Request validation failed",
                    "value": {
                      "detail": "A parameter failed validation, including declared caps. `errors[]` names each offending pointer. Caps are refused, never silently clamped.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Request validation failed",
                      "type": "/v1/errors/validation_failed"
                    }
                  }
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "service_degraded": {
                    "summary": "Service degraded",
                    "value": {
                      "detail": "A required store is unavailable.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 503,
                      "title": "Service degraded",
                      "type": "/v1/errors/service_degraded"
                    }
                  }
                }
              }
            },
            "description": "Service degraded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List glossary terms",
        "tags": [
          "glossary"
        ],
        "x-glasswell-dataset": {
          "anchors": [],
          "collection_pointer": "",
          "columns": {
            "default": [
              "/term_id",
              "/term",
              "/short_definition",
              "/domain_tags",
              "/aliases"
            ],
            "hidden": [],
            "hidden_reason": {},
            "sort": "/term"
          },
          "detail_operation": "get_glossary_term",
          "facets": [
            "q",
            "domain_tag"
          ],
          "group": "vocabulary",
          "id": "glossary",
          "intro": "nb_dataset_glossary",
          "order": 40,
          "row_id": [
            "/term_id"
          ],
          "title": "Glossary"
        },
        "x-glasswell-request-example": {
          "path": {},
          "query": {
            "limit": 5
          }
        }
      }
    },
    "/v1/glossary/index": {
      "get": {
        "description": "The compact artifact the UI highlighter consumes: pre-lowercased surface forms expanded from terms and aliases, each with its word count, plus the stopword list. Terms flagged non-highlightable are served as stopwords so a glossary containing ordinary words does not underline every third word on the page. Declared before /glossary/{term} so `index` is not matched as a term.",
        "operationId": "get_glossary_index",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeModel_GlossaryIndex_"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "service_degraded": {
                    "summary": "Service degraded",
                    "value": {
                      "detail": "A required store is unavailable.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 503,
                      "title": "Service degraded",
                      "type": "/v1/errors/service_degraded"
                    }
                  }
                }
              }
            },
            "description": "Service degraded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Term index for the highlighter",
        "tags": [
          "glossary"
        ],
        "x-glasswell-request-example": {
          "path": {},
          "query": {}
        }
      }
    },
    "/v1/glossary/{term}": {
      "get": {
        "description": "Accepts either the `gt_*` id or the surface form, case-folded: an agent holding a meta.labels value has an id and a human typing a URL has a word. Returns the expanded definition, related terms, the sources the definition came from, and the API fields that carry the term.",
        "operationId": "get_glossary_term",
        "parameters": [
          {
            "description": "Term id (gt_*) or the surface form.",
            "in": "path",
            "name": "term",
            "required": true,
            "schema": {
              "description": "Term id (gt_*) or the surface form.",
              "title": "Term",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeModel_GlossaryTermDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "not_found": {
                    "summary": "Not found",
                    "value": {
                      "detail": "The resource does not exist, or exists but is not visible to this principal. Both answer the same way so probing reveals nothing.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 404,
                      "title": "Not found",
                      "type": "/v1/errors/not_found"
                    }
                  }
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "service_degraded": {
                    "summary": "Service degraded",
                    "value": {
                      "detail": "A required store is unavailable.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 503,
                      "title": "Service degraded",
                      "type": "/v1/errors/service_degraded"
                    }
                  }
                }
              }
            },
            "description": "Service degraded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "One glossary term",
        "tags": [
          "glossary"
        ],
        "x-glasswell-request-example": {
          "path": {
            "term": "gt_report_vintage"
          },
          "query": {}
        }
      }
    },
...
  }
}